### 2024-03-26 #6543 @editorialbot
[PRE REVIEW]: Salt: Multimodal Multitask Machine Learning for High Energy Physics
**Submitting author:** @samvanstroud (<a href="http://orcid.org/0000-0002-7969-0301">Samuel Van Stroud</a>)
**Repository:** https://github.com/umami-hep/salt
**Branch with paper.md** (empty if default branch):
**Version:** v0.5
