### 2024-03-26 #6543 @kyleniemeyer
Hello @samvanstroud, thanks for your submission to JOSS. I noticed that the GitHub repo submitted is just a mirror of the repo at https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/salt - is that the actual repo for the software? We'd prefer to link the paper with the official repo, where people would go to submit issues, etc.
