### 2024-03-26 #6543 @samvanstroud
Hi @kyleniemeyer, thanks for starting the review! Yes the GitLab repo is where most of the development takes place, but unfortunately it requires a sign-in so doesn't meet the [JOSS submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements), hence the mirror.

For what it's worth, we are following the same approach as for a previous submission: https://github.com/openjournals/joss-reviews/issues/5833.
