### 2024-04-29 d1c0af1
Extend featurewise transformations to layers in encoder
