### 2024-05-09 49198e5
Tweaks to Transformerv2: layerscale + varlen_attn + small optim
