+++ salt/models/transformer_v2.py
@@ -9,57 +9,61 @@ Features:
 - RMSNorm https://arxiv.org/abs/1910.07467
 """
 
-from abc import ABC
+import warnings
+from functools import partial
 
 import torch
-from torch import BoolTensor, Size, Tensor, nn
+import torch.nn.functional as F
+from torch import BoolTensor, Tensor, nn
 
 import salt.models.layernorm as layernorms
 from salt.stypes import Tensors
+from salt.utils.tensor_utils import redo_padding, undo_padding
 
 
 def merge_masks(
-    q_mask: BoolTensor | None,
     kv_mask: BoolTensor | None,
     attn_mask: BoolTensor | None,
-    q_shape: Size,
-    k_shape: Size,
-) -> BoolTensor:
+    q_shape: Tensor,
+) -> BoolTensor | None:
     """Create a full attention mask which incorporates the padding information.
 
-    Using pytorch transformer convention:
+    Using pytorch transformer convention for padding
         False: Real node
         True:  Zero padded
 
+    Using pytorch transformer convention for attention mask
+        False:  Not allowed in attention mechanism
+        True:   Allowed in attention mechanism
+
+    Designing attention mask such that padded tokens can't send information.
+    But they can receive them.
+    This prevents Nans in the attention scores caused by the softmax
+
     Parameters
     ----------
-    q_mask : BoolTensor | None
-        Mask for the queries, of shape (batch, q_len).
     kv_mask : BoolTensor | None
         Mask for the keys and values, of shape (batch, kv_len).
     attn_mask : BoolTensor | None
         Full attention mask, of shape (batch, q_len, kv_len).
     q_shape : Size
         Shape of the queries tensor, (batch, q_len, dim).
-    k_shape : Size
-        Shape of the keys tensor, (batch, kv_len, dim).
     """
     # Create the full mask which combines the attention and padding masks
     mask = None
 
-    # if both masks exist, combine them
-    if q_mask is not None and kv_mask is not None:
-        mask = q_mask.unsqueeze(-1) | kv_mask.unsqueeze(-2)
-
-    # if only one mask exists, expand it to the other dimension
-    if q_mask is None and kv_mask is not None:
+    # if the kv_mask mask exists, ensure that padded tokens never send information
+    if kv_mask is not None:
         mask = kv_mask.unsqueeze(-2).expand(-1, q_shape[-2], -1)
-    if kv_mask is None and q_mask is not None:
-        mask = q_mask.unsqueeze(-1).expand(-1, -1, k_shape[-2])
+        mask = ~mask  # convert the mask such that True is a valid token
 
     # include the attention mask
     if attn_mask is not None:
-        mask = attn_mask if mask is None else attn_mask | mask
+        mask = attn_mask if mask is None else attn_mask & mask
+
+    # Unsqueeze the mask to give it a dimension for num_head broadcasting
+    if mask is not None:
+        mask = mask.unsqueeze(1)
 
     return mask
 
@@ -70,50 +74,88 @@ def repeat_kv(keys: Tensor, values: Tensor, repeats: int, dim: int):
     return keys, values
 
 
-def torch_meff_attn(q: Tensor, k: Tensor, v: Tensor, mask: BoolTensor, dropout: float) -> Tensor:
-    # masking can lead to nans, see
-    # - https://github.com/pytorch/pytorch/issues/110213
-    # - https://github.com/pytorch/pytorch/issues/103749
-    # to get round this, can transform the mask from a bool to float
-    # mask = (1.0 - mask.to(q.dtype)) * torch.finfo(q.dtype).min
-    # but don't need this if add_zero_attn is True
+def change_attn_backends(module: nn.Module, backend: str) -> None:
+    """Recursively change the attention backend of a module and all its children.
 
-    # TODO: change mask convention
[…]
