+++ docs/api/transformer.md
@@ -2,7 +2,6 @@
 ## ::: salt.models.transformer_v2.Attention
     options:
       members: [forward]
-## ::: salt.models.transformer_v2.SelfAttention
 ## ::: salt.models.transformer_v2.GLU
 ## ::: salt.models.transformer_v2.EncoderLayer
 ## ::: salt.models.transformer_v2.TransformerV2
