### 2024-05-09 d29faaf
Minor follow up to transformer changes
