+++ salt/configs/GN3.yaml
@@ -11,7 +11,6 @@ model:
   model:
     class_path: salt.models.SaltModel
     init_args:
-      num_register_tokens: 0 # Registers have moved to the encoder!
 
       init_nets:
         - input_name: tracks
@@ -26,9 +25,8 @@ model:
           num_layers: 6
           embed_dim: *embed_dim
           out_dim: &out_dim 128
+          attn_type: flash-varlen
           norm: LayerNorm
-          attn_type: torch-math
-          do_final_norm: true
           ls_init: 1.0e-2
           dense_kwargs:
             activation: SiLU
