### 2024-06-20 e1f33a1
Ensure class_dict is not null before logging cd_path
