+++ salt/to_onnx.py
@@ -16,7 +16,7 @@ from tqdm import tqdm
 from salt.models.task import mask_fill_flattened
 from salt.models.transformer_v2 import change_attn_backends
 from salt.modelwrapper import ModelWrapper
-from salt.utils.inputs import inputs_sep_no_pad, inputs_sep_with_pad
+from salt.utils.inputs import inputs_sep_with_pad_multi_sequece
 from salt.utils.union_find import get_node_assignment_jit
 
 torch.manual_seed(42)
@@ -94,19 +94,34 @@ def get_probs(outputs: Tensor):
 
 
 class ONNXModel(ModelWrapper):
-    def __init__(self, name: str | None = None, include_aux: bool = False, **kwargs) -> None:
+    def __init__(
+        self, name: str | None = None, include_aux: bool = False, onnx_feature_map=None, **kwargs
+    ) -> None:
         super().__init__(**kwargs)
         self.name = name if name else self.name
-        assert len(self.model.init_nets) == 1, "Multi input ONNX models are not yet supported."
         assert "_" not in self.name, "Model name cannot contain underscores."
         assert "-" not in self.name, "Model name cannot contain dashes."
         self.include_aux = include_aux
-        self.const = "tracks"
-        self.input_names = ["jet_features", "track_features"]
-        jets, tracks = inputs_sep_no_pad(
-            1, 40, self.input_dims[self.global_object], self.input_dims[self.const]
-        )
-        self.example_input_array = jets, tracks.squeeze(0)  # used for the tracing during export
+        self.feature_map = onnx_feature_map
+        self.aux_sequence_object = "tracks"
+
+        example_input_list = []
+        self.salt_names = []
+        self.input_names = []
+        self.aux_sequence_index = 1
+
+        for i, feature in enumerate(self.feature_map):
+            if feature["name_salt"] == self.global_object:
+                example_input_list.append(torch.rand(1, self.input_dims[self.global_object]))
+            else:
+                example_input_list.append(
+                    torch.rand(1, 40, self.input_dims[feature["name_salt"]]).squeeze(0)
+                )
+            if feature["name_salt"] == self.aux_sequence_object:
+                self.aux_sequence_index = i
+            self.salt_names.append(feature["name_salt"])
+            self.input_names.append(feature["name_athena_out"])
+        self.example_input_array = tuple(example_input_list)
 
     @property
     def model_name(self) -> str:
@@ -138,7 +153,10 @@ class ONNXModel(ModelWrapper):
     def dynamic_axes(self) -> dict[str, dict[int, str]]:
         """Let ONNX know which inputs/outputs have dynamic shape (i.e. can vary in length)."""
         # dynamic inputs
-        dynamic_axes = {"track_features": {0: "n_tracks"}}
+        dynamic_axes = {}
+        for feature in self.feature_map:
+            if not feature["is_global"]:
+                dynamic_axes.update({feature["name_athena_out"]: {0: feature["athena_num_name"]}})
 
         # dynamic outputs
         if self.include_aux:
@@ -150,14 +168,21 @@ class ONNXModel(ModelWrapper):
                 dynamic_axes[out_name] = {0: "n_tracks"}
         return dynamic_axes
 
-    def forward(self, jets: Tensor, tracks: Tensor, labels=None):  # type: ignore[override]
-        _ = labels
-
+    def forward(self, *args):  # type: ignore[override]
+        """Forward pass through the model."""
+        """ the arguments should be passed in the same order they appear in the feature map """
[…]
