### 2024-06-27 203180f
General options to submit_slurm.py
