### 2024-07-30 2dfc10d
MF Onnx and cleaner output predictions
