+++ salt/to_onnx.py
@@ -13,10 +13,14 @@ from torch import Tensor
 from torch.nn.functional import softmax
 from tqdm import tqdm
 
+from salt.models.maskformer import get_maskformer_outputs
 from salt.models.task import mask_fill_flattened
 from salt.models.transformer_v2 import change_attn_backends
 from salt.modelwrapper import ModelWrapper
-from salt.utils.inputs import inputs_sep_with_pad_multi_sequece
+from salt.utils.configs import MaskformerConfig
+from salt.utils.inputs import (
+    inputs_sep_with_pad_multi_sequece,
+)
 from salt.utils.union_find import get_node_assignment_jit
 
 torch.manual_seed(42)
@@ -79,6 +83,10 @@ def parse_args(args):
         action="store_true",
     )
     parser.add_argument(
+        "-mf",
+        "--object_name",
+    )
+    parser.add_argument(
         "-f",
         "--force",
         help="Run with uncomitted changes.",
@@ -95,13 +103,33 @@ def get_probs(outputs: Tensor):
 
 class ONNXModel(ModelWrapper):
     def __init__(
-        self, name: str | None = None, include_aux: bool = False, onnx_feature_map=None, **kwargs
+        self,
+        onnx_feature_map: list[dict],
+        name: str | None = None,
+        include_aux: bool = False,
+        object_name: str | None = None,
+        mf_config: dict | None = None,
+        **kwargs,
     ) -> None:
         super().__init__(**kwargs)
         self.name = name if name else self.name
         assert "_" not in self.name, "Model name cannot contain underscores."
         assert "-" not in self.name, "Model name cannot contain dashes."
         self.include_aux = include_aux
+
+        if sum([bool(object_name), bool(mf_config)]) not in {0, 2}:
+            raise ValueError("If one of object name or mf config is defined, so must the other.")
+        self.object = object_name
+        self.mf_config = MaskformerConfig(**mf_config) if mf_config else None
+        if self.object and self.mf_config:
+            self.object_params = {
+                "class_label": self.mf_config.object.class_label,
+                "label_map": [f"p{name}" for name in self.mf_config.object.class_names],
+            }
+
+        self.has_global_task = (
+            len([t for t in self.model.tasks if t.input_name == self.global_object]) > 0
+        )
         self.feature_map = onnx_feature_map
         self.aux_sequence_object = "tracks"
 
@@ -133,10 +161,12 @@ class ONNXModel(ModelWrapper):
         """The output names are a list of strings, one for each output of the model."""
         # get the global task output
         global_tasks = [t for t in self.model.tasks if t.input_name == self.global_object]
-        assert len(global_tasks) == 1, "Multi global task ONNX models are not yet supported."
-        object_classes = global_tasks[0].class_names
-        outputs = [f"{self.model_name}_p{flav.rstrip('jets')}" for flav in object_classes]
-
+        assert len(global_tasks) <= 1, "Multi global task ONNX models are not yet supported."
+        if self.has_global_task:
+            object_classes = global_tasks[0].class_names
+            outputs = [f"{self.model_name}_p{flav.rstrip('jets')}" for flav in object_classes]
+        else:
+            outputs = []
         # aux task output names
         if self.include_aux:
             if "track_origin" in [t.name for t in self.model.tasks]:
@@ -146,6 +176,16 @@ class ONNXModel(ModelWrapper):
             if "track_vertexing" in [t.name for t in self.model.tasks]:
                 out_name = f"{self.model_name}_VertexIndex"
                 outputs.append(out_name)
+        if self.object:
+            regression_task = [
[…]
