### 2024-08-19 #6543 @kyleniemeyer
@samvanstroud the issue is actually finding a JOSS editor, and not reviewers. We unfortunately have very low editor availability right now.
