### 2024-08-22 3cce707
Fixed bug in remapping labels for classification
