### 2024-09-02 fb66c6b
Fix NodeQueryGAP
