### 2024-09-02 1866163
Batch pred writer
