+++ salt/callbacks/predictionwriter.py
@@ -1,9 +1,9 @@
 from collections import defaultdict
 from pathlib import Path
 
-import h5py
 import numpy as np
 from ftag import Flavours as Flavs
+from ftag.hdf5 import H5Writer
 from lightning import Callback, LightningModule, Trainer
 from numpy.lib.recfunctions import unstructured_to_structured as u2s
 
@@ -66,9 +66,11 @@ class PredictionWriter(Callback):
         if stage != "test":
             return
 
+        self.writer = None
         # some basic info
         self.trainer = trainer
         self.ds = trainer.datamodule.test_dataloader().dataset
+        self.batch_size = trainer.datamodule.batch_size
         self.test_suff = trainer.datamodule.test_suff
         self.file = self.ds.file
         self.num = len(self.ds)
@@ -133,10 +135,60 @@ class PredictionWriter(Callback):
         suffix = f"_{self.test_suff}" if self.test_suff is not None else ""
         return Path(out_dir / f"{out_basename}__test_{sample}{suffix}.h5")
 
+    def _write_batch_outputs(self, batch_outputs, pad_masks, batch_idx):
+        to_write = {}
+        blow = batch_idx * self.batch_size
+        bhigh = (batch_idx + 1) * self.batch_size
+        for input_name, outputs in batch_outputs.items():
+            this_outputs = []
+
+            if input_name in self.input_map:
+                name = self.input_map[input_name]
+
+                # get input variables
+                input_variables = self.extra_vars[name]
+                if not input_variables:
+                    input_variables = self.file[name].dtype.names
+                if name in self.file:
+                    inputs = self.file[name].fields(input_variables)[blow:bhigh]
+                    this_outputs.append(inputs)
+            else:
+                name = input_name
+                inputs = None
+
+            for preds in outputs.values():
+                if inputs is not None:
+                    this_outputs.append(maybe_pad(preds, inputs))
+                else:
+                    this_outputs.append(preds)
+
+            # add mask if present
+            if name in pad_masks:
+                pad_mask = pad_masks[name].cpu()
+                pad_mask = u2s(np.expand_dims(pad_mask, -1), dtype=np.dtype([("mask", "?")]))
+                this_outputs.append(maybe_pad(pad_mask, inputs))
+
+            to_write[name] = join_structured_arrays(this_outputs)
+
+        # If the writer hasn't been created yet, create it now that we have the dtypes and shapes
+        if self.writer is None:
+            dtypes = {k: v.dtype for k, v in to_write.items()}
+            shapes = {k: (self.num,) + v.shape[1:] for k, v in to_write.items()}
+            self.writer = H5Writer(
+                dst=self.output_path,
+                dtypes=dtypes,
+                shapes=shapes,
+                shuffle=False,
+                precision="half" if self.half_precision else "full",
+            )
+        self.writer.write(to_write)
+
     def on_test_batch_end(self, trainer, module, outputs, batch, batch_idx):  # noqa: ARG002
         preds = outputs
         _, pad_masks, labels = batch
         add_mask = False
+        to_write: dict = {input_name: {} for input_name in {t.input_name for t in self.tasks}}
+        out_pads = {}
         for task in self.tasks:
             if (
                 (not self.write_tracks and task.input_name == "tracks")
@@ -164,107 +216,56 @@ class PredictionWriter(Callback):
                 if f"{module.name}_" not in task.name:
                     task.name = f"{module.name}_" + task.name
                 means, stddevs = task.run_inference(this_preds, labels, self.precision)
-            if task.name not in self.outputs[task.input_name]:
-                self.outputs[task.input_name][task.name] = []
+            if task.name not in to_write[task.input_name]:
[…]
