+++ mkdocs.yaml
@@ -45,6 +45,7 @@ nav:
       - api/transformer.md
       - api/models.md
       - api/tasks.md
+  - test: test/test.md
 
 plugins:
   - search
