### 2024-09-07 #6543 @arfon
@samvanstroud – could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?
