### 2024-09-07 #6543 @samvanstroud
Hi @arfon - thanks so much for taking this on! We really appreciate it.

I've just had a look, here are some candidates (in no particular order):

- https://reviewers.joss.theoj.org/reviewers/3050
- https://reviewers.joss.theoj.org/reviewers/2718
- https://reviewers.joss.theoj.org/reviewers/3558
- https://reviewers.joss.theoj.org/reviewers/3363
- https://reviewers.joss.theoj.org/reviewers/4532
- https://reviewers.joss.theoj.org/reviewers/1907
- https://reviewers.joss.theoj.org/reviewers/4088
