### 2024-09-09 ca2dea9
Cleanup Tasks, PredictionWriter, and to_onnx.py
