+++ salt/models/task.py
@@ -3,10 +3,12 @@ from collections.abc import Mapping
 
 import numpy as np
 import torch
+from ftag import Flavours
 from numpy.lib.recfunctions import unstructured_to_structured as u2s
 from torch import Tensor, nn
 
 from salt.models import Dense
+from salt.stypes import Tensors
 from salt.utils.array_utils import listify
 from salt.utils.class_names import CLASS_NAMES
 from salt.utils.scalers import RegressionTargetScaler
@@ -109,6 +111,12 @@ class ClassificationTask(TaskBase):
             )
         self.use_class_dict = use_class_dict
 
+    @property
+    def output_names(self) -> list[str]:
+        assert self.class_names is not None
+        pxs = [f"{Flavours[c].px}" if c in Flavours else f"p{c}" for c in self.class_names]
+        return [f"{self.model_name}_{px}" for px in pxs]
+
     def apply_sample_weight(self, loss: Tensor, labels_dict: Mapping) -> Tensor:
         """Apply per sample weights, if specified."""
         if self.sample_weight is None:
@@ -159,16 +167,23 @@ class ClassificationTask(TaskBase):
 
         return preds, loss
 
-    def run_inference(self, preds: Tensor, pad_mask: Tensor | None = None, precision: str = "f4"):
+    def run_inference(self, preds: Tensor, pad_mask: Tensor | None = None) -> Tensor:
         if pad_mask is None:
             assert preds.ndim == 2
             probs = torch.softmax(preds, dim=-1)
         else:
             assert preds.ndim == 3
             probs = masked_softmax(preds, pad_mask.unsqueeze(-1))
-        assert self.class_names is not None
-        dtype = np.dtype([(n, precision) for n in self.class_names])
+        return probs
+
+    def get_h5(self, preds: Tensor, pad_mask: Tensor | None = None) -> np.ndarray:
+        probs = self.run_inference(preds, pad_mask)
+        dtype = np.dtype([(n, "f4") for n in self.output_names])
         return u2s(probs.float().cpu().numpy(), dtype)
+
+    def get_onnx(self, preds: Tensor, pad_mask: Tensor | None = None) -> tuple:
+        probs = self.run_inference(preds, pad_mask)
+        return tuple(output.squeeze() for output in torch.split(probs, 1, -1))
 
 
 class RegressionTaskBase(TaskBase, ABC):
@@ -312,6 +327,10 @@ class RegressionTask(RegressionTaskBase):
             )
         self.scaler = scaler
 
+    @property
+    def output_names(self) -> list[str]:
+        return [f"{self.model_name}_{self.name}_{x}" for x in self.targets]
+
     def forward(
         self,
         x: Tensor,
@@ -335,19 +354,22 @@ class RegressionTask(RegressionTaskBase):
 
         return preds, loss
 
-    def run_inference(self, preds: Tensor, targets_dict: Mapping, precision: str = "f4"):
+    def run_inference(self, preds: Tensor, labels: Tensors) -> Tensor:
         if self.target_denominators is not None:
             for i in range(len(self.targets)):
-                preds[:, i] = (
-                    preds[:, i] * targets_dict[self.input_name][self.target_denominators[i]]
-                )
+                preds[:, i] *= labels[self.input_name][self.target_denominators[i]]
         elif self.norm_params is not None:
             for i in range(len(self.norm_params["mean"])):
-                preds[:, i] = preds[:, i] * self.norm_params["std"][i] + self.norm_params["mean"][i]
+                preds[:, i] *= self.norm_params["std"][i]
+                preds[:, i] += self.norm_params["mean"][i]
         elif self.scaler is not None:
[…]
