+++ requirements.txt
@@ -16,7 +16,7 @@ jsonargparse[all]==4.32.0
 torchmetrics==1.2.1
 onnx==1.16.0
 onnxruntime==1.15.1
-atlas-ftag-tools==0.1.18
+atlas-ftag-tools==0.2.2
 scipy==1.12.0
 
 # S3 and muTransfer
