### 2024-09-09 #6543 @arfon
:wave: @rudraprsd @tkoyama010 @r2stanton @AstroBarker @divijghose would any of you be willing to review this submission for JOSS? The submission under consideration is Salt: Multimodal Multitask Machine Learning for High Energy Physics

Many thanks
Arfon
