### 2024-09-12 62e2c83
nan check in compare models script
