### 2024-09-13 #7217 @tkoyama010
The documentation shows how to install this library directly from the repository. Is there a way to install from PyPI and conda-forge?
