### 2024-09-13 #7217 @samvanstroud
Hi @tkoyama010! We don't currently offer this since there hasn't been a clear need from users, but it's certainly something we could add if you think it would be useful. Thanks
