### 2024-09-17 #7217 @tkoyama010
> Hi @tkoyama010! We don't currently offer this since there hasn't been a clear need from users, but it's certainly something we could add if you think it would be useful. Thanks

Awesome. I am registering with them is recommended as it lowers the psychological hurdle for users.
