### 2024-09-19 5908b11
Adding Base Docker Image
