+++ setup/docker/testing/Dockerfile
@@ -1,5 +1,5 @@
-# base image
-FROM pytorch/pytorch:2.2.1-cuda12.1-cudnn8-devel
+# Using base image from pytorch
+FROM ${BASE_IMAGE}
 
 # local and envs
 ENV LANG C.UTF-8
@@ -7,10 +7,10 @@ ENV LC_ALL C.UTF-8
 ENV PIP_ROOT_USER_ACTION=ignore
 ENV PIP_NO_CACHE_DIR=false
 ARG DEBIAN_FRONTEND=noninteractive
-WORKDIR /salt
 
 # add some packages
-RUN apt-get update && apt-get install -y git h5utils wget vim build-essential jq
+RUN apt-get update && \
+    apt-get install -y git h5utils wget vim nano build-essential jq
 
 # update python pip
 RUN python -m pip install --upgrade pip
@@ -28,7 +28,9 @@ RUN python -m pip install puma-hep umami-preprocessing
 RUN python -m pip install wheel packaging ninja
 RUN python -m pip install flash-attn==2.5.7
 
-# copy and install package
-COPY . .
-RUN python -m pip install -e .
+# Create folder and copy salt
+COPY . /salt
+WORKDIR /salt
 
+# Install salt
+RUN cd /salt && python -m pip install .
