+++ .gitlab-ci.yml
@@ -3,9 +3,11 @@ workflow:
     - if: '$CI_PIPELINE_SOURCE == "merge_request_event"'
       when: never
     - when: always
+
 stages:
   - lint
-  - builds
+  - base_image_build
+  - salt_image_build
   - tests
   - coverage
   - docs
@@ -16,3 +18,6 @@ include:
   - .gitlab/.ci-test.yaml
   - .gitlab/.ci-coverage.yaml
   - .gitlab/.ci-docs.yaml
+
+variables:
+  PYTORCHTAG: 2.2.1-cuda12.1-cudnn8-devel
