### 2024-09-19 f719f77
Revert "Merge branch 'alfroch-add-base-image' into 'main'"
