+++ .gitlab/.ci-publish.yaml
@@ -1,0 +1,16 @@
+stages:
+  - builds
+  - publish
+
+deploy_pypi:
+  stage: publish
+  image: python:3.11-slim
+  needs: []
+  before_script:
+    - python -m pip install --upgrade pip setuptools wheel twine
+  script:
+    - python -m pip install build
+    - python -m build
+    - twine upload dist/* --username $TWINE_USERNAME --password $TWINE_PASSWORD
+  rules:
+    - if: $CI_COMMIT_TAG && $CI_COMMIT_BRANCH == $CI_DEFAULT_BRANCH && $CI_PROJECT_PATH=="atlas-flavor-tagging-tools/algorithms/salt"
