### 2024-09-19 38a6958
update publish rule
