### 2024-09-20 e45216b
Cleanup packaging and tests
