+++ setup/install_flash.sh
@@ -1,5 +1,0 @@
-# To optionally install the flash-attn package for transformer-v2 models
-python -m pip install wheel packaging
-python -m pip install ninja==1.11.1.1
-python -m pip install flash-attn==2.5.7
-
