+++ requirements-flash.txt
@@ -1,0 +1,5 @@
+# FlashAttention
+wheel
+packaging
+ninja==1.11.1.1
+flash-attn==2.5.7
