+++ pyproject.toml
@@ -20,7 +20,7 @@ setup_muP = "salt.utils.muP_utils.main_muP:main"
 download_S3 = "salt.utils.file_utils:download_from_S3"
 
 [tool.setuptools]
-packages = ["salt"]
+packages = { find = { include = ["salt*"] } }
 include-package-data = true
 
 [tool.setuptools.dynamic]
@@ -40,11 +40,58 @@ line-length = 100
 preview = true
 lint.select = ["ALL", "D212", "D417"]
 lint.ignore = [
-    "COM", "D100", "D101", "D102", "D103", "D104", "D105", "D205", "D401", "EM", "FIX", "FBT",
-    "S101", "S404", "S602", "PLR2004", "PLR0912", "PLR0913", "PLR0914", "PLR0915", "PLR0917",
-    "PLC0415", "G004", "PD901", "N802", "C901", "DTZ005", "DTZ007", "INP", "EXE002", "TD002", 
-    "ANN001", "ANN002", "ANN003", "ANN101", "ANN201", "ANN202", "ANN204", "CPY001", "TRY003", "PLR1702",
-    "N803", "N806", "N812", "T201", "PLW1514", "PTH123", "RUF015", "RUF017", "PLR6301", "ERA", "ISC001", 
+    "COM",
+    "D100",
+    "D101",
+    "D102",
+    "D103",
+    "D104",
+    "D105",
+    "D205",
+    "D401",
+    "EM",
+    "FIX",
+    "FBT",
+    "S101",
+    "S404",
+    "S602",
+    "PLR2004",
+    "PLR0912",
+    "PLR0913",
+    "PLR0914",
+    "PLR0915",
+    "PLR0917",
+    "PLC0415",
+    "G004",
+    "PD901",
+    "N802",
+    "C901",
+    "DTZ005",
+    "DTZ007",
+    "INP",
+    "EXE002",
+    "TD002",
+    "ANN001",
+    "ANN002",
+    "ANN003",
+    "ANN101",
+    "ANN201",
+    "ANN202",
+    "ANN204",
+    "CPY001",
+    "TRY003",
+    "PLR1702",
+    "N803",
+    "N806",
+    "N812",
+    "T201",
+    "PLW1514",
+    "PTH123",
+    "RUF015",
+    "RUF017",
+    "PLR6301",
+    "ERA",
+    "ISC001",
 ]
 
 [tool.ruff.lint.pydocstyle]
