### 2024-09-23 86b6592
Update GN3 config for GLS loss and lion optimiser
