+++ requirements.txt
@@ -18,3 +18,4 @@ onnx==1.16.0
 onnxruntime==1.15.1
 atlas-ftag-tools==0.2.4
 scipy==1.12.0
+lion-pytorch==0.2.2
