### 2024-09-23 b62a8bf
Remove check on model parameters
