### 2024-09-23 #7217 @samvanstroud
Thanks @tkoyama010. It's now available on PyPI: https://pypi.org/project/salt-ml/
