### 2024-09-24 ed9e1b4
Moving GN3 config and use input map for reading class dict
