### 2024-10-01 #7217 @tkoyama010
Last week, I was diagnosed with ulcerative colitis by a doctor. I will be taking some time to recover, so my responses will be significantly delayed. I will return to the community when I am recover.
If it is not possible to extend the review deadline, I would be very sorry, but I would like to be removed from the reviewer list.
