### 2024-10-02 a00e8cb
Don't write tracks by default
