### 2024-10-08 d140a60
Cleaning up gn3_dev configs
