### 2024-10-14 #6543 @samvanstroud
@arfon how about https://reviewers.joss.theoj.org/reviewers/3050 or https://reviewers.joss.theoj.org/reviewers/3558?
