### 2024-10-22 e7598c9
Config for GN3 track/source type label training
