+++ salt/configs/GN3_dev/GN3_baseline_loose.yaml
@@ -1,3 +1,5 @@
+# To be use on top of
+# GN3_baseline.yaml
 name: GN3_baseline_loose
 
 data:
