### 2024-10-24 1c47fb2
Supporting RegressionTask in ONNX export & multiple global tasks
