### 2024-10-26 #6543 @arfon
:wave: @GarrettMerz – of you be willing to review this submission for JOSS? The submission under consideration is Salt: Multimodal Multitask Machine Learning for High Energy Physics

Many thanks
Arfon
