### 2024-10-29 5f53b76
GN3v00 config
