### 2024-11-01 #7217 @arfon
@pikarpov-lanl @asogaard – would you be willing to review this submission for JOSS? The submission under consideration is _Salt: Multimodal Multitask Machine Learning for High Energy Physics_

Based on your experience, and past papers published in JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
