### 2024-11-01 #7217 @asogaard
Hi @arfon,
Thanks for considering me for reviewing this paper. I very much appreciated the JOSS review  process myself. However, as I have left academia, I think you'd be better off finding a reviewer who is more active in the field and can provide a better review.
Best,
Andreas
