### 2024-11-17 umami-hep/salt#1 @divijghose
Typos and corrections in the paper
Hello,

After reviewing the paper for JOSS, I have found the following typos:
1. Line 19: "configurabiltiy" should be "configurability"
2. Line 50: "admist" should be "amidst"
3. Line 63: "concatentated" should be "concatenated"

Could you please look into this, thanks!
