### 2024-11-17 #7217 @divijghose
Hello @samvanstroud,

I have finished reviewing the paper. It is well written, but I have found a few typos. I have raised an issue here: https://github.com/umami-hep/salt/issues/1#issue-2665773370

Could you please correct these, thank you.
