### 2024-11-18 umami-hep/salt#1 @samvanstroud
Hi @divijghose, many thanks for the review! The typos have been fixed in https://github.com/umami-hep/salt/commit/3a5093965d14ab653bf34280b860daa43addb205
