### 2024-11-18 #7217 @samvanstroud
Hi @divijghose, many thanks for taking a look. The typos have been fixed in https://github.com/umami-hep/salt/commit/3a5093965d14ab653bf34280b860daa43addb205.
