+++ salt/models/transformer_v2.py
@@ -9,6 +9,7 @@ Features:
 - RMSNorm https://arxiv.org/abs/1910.07467
 """
 
+import math
 import warnings
 from functools import partial
 
@@ -161,6 +162,8 @@ class Attention(nn.Module):
         attn_type: str = "torch-meff",
         dropout: float = 0.0,
         bias: bool = True,
+        diff_attention: bool = False,
+        depth: int = 1,
     ) -> None:
         """Multihead attention module.
 
@@ -176,6 +179,10 @@ class Attention(nn.Module):
             Dropout rate.
         bias : bool, optional
             Whether to include bias terms.
+        diff_attention : bool, optional
+            Use differential attention or not
+        depth : int, optional
+            Number of current attention layer
         """
         super().__init__()
         assert embed_dim % num_heads == 0, "Dim not div by the number of heads!"
@@ -188,6 +195,26 @@ class Attention(nn.Module):
         self.dropout = dropout
         self.bias = bias
         self.attn_type = attn_type
+        self.diff_attention = diff_attention
+        self.depth = depth
+
+        if self.diff_attention:
+            self.head_dim = self.head_dim // 2
+            self.lambda_init = 0.8 - 0.6 * math.exp(-0.3 * depth)
+            self.lambda_q1 = nn.Parameter(
+                torch.zeros(self.head_dim, dtype=torch.float32).normal_(mean=0, std=0.1)
+            )
+            self.lambda_k1 = nn.Parameter(
+                torch.zeros(self.head_dim, dtype=torch.float32).normal_(mean=0, std=0.1)
+            )
+            self.lambda_q2 = nn.Parameter(
+                torch.zeros(self.head_dim, dtype=torch.float32).normal_(mean=0, std=0.1)
+            )
+            self.lambda_k2 = nn.Parameter(
+                torch.zeros(self.head_dim, dtype=torch.float32).normal_(mean=0, std=0.1)
+            )
+
+            self.subln = layernorms.RMSNorm(2 * self.head_dim)
 
         # Better parallelism for self-attention when using parameters directly
         self.in_proj_weight = nn.Parameter(torch.empty(3 * embed_dim, embed_dim))
@@ -227,6 +254,18 @@ class Attention(nn.Module):
             nn.init.constant_(self.in_proj_bias, 0.0)
         self.out_proj.reset_parameters()
 
+    def _weight_by_lambda(self, attn1: Tensor, attn2: Tensor) -> Tensor:
+        lambda_1 = torch.exp(torch.sum(self.lambda_q1 * self.lambda_k1, dim=-1).float()).type_as(
+            attn1
+        )
+        lambda_2 = torch.exp(torch.sum(self.lambda_q2 * self.lambda_k2, dim=-1).float()).type_as(
+            attn1
+        )
+        lambda_full = lambda_1 - lambda_2 + self.lambda_init
+        attn = attn1 - lambda_full * attn2
+
+        return self.subln(attn) * (1 - self.lambda_init)
+
     def _flash_forward(self, x: Tensor, culens: Tensor, maxlen: int) -> Tensor:
         """FlashAttention backend."""
         # Perform the packed input projection
@@ -236,6 +275,37 @@ class Attention(nn.Module):
         # Run the flash-varlen backend
         dropout = self.dropout if self.training else 0.0
         a_out = self._flash_attn(qkv, culens, maxlen, dropout)
+        a_out = a_out.reshape(-1, self.embed_dim)
+
+        # Mix with final linear layer
[…]
