### 2024-11-24 #7217 @divijghose
@samvanstroud and co-authors, thank you for this interesting and useful library. While the documentation is exhaustive and follows best practices, I have a few recommendations:
- **Statement of need** : Could you add a short statement of need in the documentation homepage? This would help first-time users, especially those outside your organization. This could be a compressed version of the statement of need that you have added in the software paper.
- **Contributing** : As mentioned in the documentation, users can seek support by opening an issue in the repository or on mattermost. Since there is support for third party users outside your organization via the repository, I have ticked the relevant section in the checklist. But I was wondering if there was any mechanism to bring important issues discussed on mattermost to the repo/documentation, if they have been missed by general users. I believe this would make things more _open_.

Let me know what you think!

Regards,
Divij
