### 2024-11-25 #7217 @samvanstroud
Hi @divijghose, thanks again for the detailed review.

- **Statement of need**: added in https://github.com/umami-hep/salt/commit/bc9602c6d77c5522568d147ae92fe654daa425cb
- **Contributing**: Thanks for bringing this point up. We encourage people to open repo issues when bugs/issues are identified. The internal discussion on MM is mostly around usage and development, so hopefully other users are not missing too much here. Will keep this in mind though, thanks 👍

Cheers,
Sam
