### 2024-12-06 1adf14e
Temporary fix for GN3 high stats training
