### 2024-12-17 77a955b
possible fix for the nan issues
