+++ salt/callbacks/__init__.py
@@ -1,13 +1,17 @@
 from salt.callbacks.checkpoint import Checkpoint
+from salt.callbacks.gradient_logger import GradientLoggerCallback
 from salt.callbacks.maskformer_metrics import MaskformerMetrics
 from salt.callbacks.performancewriter import PerformanceWriter
 from salt.callbacks.predictionwriter import PredictionWriter
 from salt.callbacks.saveconfig import SaveConfigCallback
+from salt.callbacks.weight_logger import WeightLoggerCallback
 
 __all__ = [
     "Checkpoint",
+    "GradientLoggerCallback",
     "MaskformerMetrics",
     "PerformanceWriter",
     "PredictionWriter",
     "SaveConfigCallback",
+    "WeightLoggerCallback",
 ]
