+++ salt/configs/GN3_dev/GN3_v00.yaml
@@ -56,7 +56,6 @@ model:
                 input_name: jets
                 label: flavour_label
                 use_class_dict: True
-                mask_invalid_pred: True
                 loss: torch.nn.CrossEntropyLoss
                 dense_config: &task_dense_config
                   input_size: *out_dim
@@ -70,7 +69,6 @@ model:
                 input_name: tracks
                 label: ftagTruthOriginLabel
                 use_class_dict: True
-                mask_invalid_pred: True
                 loss: torch.nn.CrossEntropyLoss
                 dense_config:
                   <<: *task_dense_config
@@ -82,7 +80,6 @@ model:
                 name: track_vertexing
                 input_name: tracks
                 label: ftagTruthVertexIndex
-                mask_invalid_pred: True
                 loss:
                   class_path: torch.nn.BCEWithLogitsLoss
                   init_args: { reduction: none }
@@ -100,7 +97,6 @@ model:
                 label: ftagTruthTypeLabel
                 label_map: {-2 : 2, -3 : 3, 5 : 4, -5 : 4, 6 : 5, -6 : 5}
                 class_names: [NoTruth, Other, Pion, Kaon, Electron, Muon]
-                mask_invalid_pred: True
                 loss:
                   class_path: torch.nn.CrossEntropyLoss
                   init_args:
@@ -117,7 +113,6 @@ model:
                 input_name: jets
                 targets: ptFromTruthDressedWZJet
                 target_denominators: pt_btagJes
-                mask_invalid_pred: True
                 loss:
                   class_path: torch.nn.L1Loss
                   init_args: { reduction: none }
@@ -133,7 +128,7 @@ data:
     flows: flows
   # loose tracks
   selections: null
-  ignore_finite_checks: True
+  non_finite_to_num: True
   variables:
     jets:
       - pt_btagJes
