### 2024-12-27 #7217 @arfon
Hi @GarrettMerz – my apologies, I missed your response over here https://github.com/openjournals/joss-reviews/issues/6543#issuecomment-2514731000
