### 2025-01-01 #7217 @divijghose
@arfon I have wrapped up my review by checking the functionality of the code. There were a few minor issues with the software paper and documentation,  which have been addressed by the authors to my satisfaction. I have no further objections.

@samvanstroud thank you for cooperation with this review, and for addressing the issues promptly. Great work, I look forward to it being published!
