+++ salt/onnx/to_onnx.py
@@ -71,14 +71,33 @@ def parse_args(args):
         action="store_true",
     )
     parser.add_argument(
-        "-a",
-        "--include_aux",
-        help="Include auxiliary task outputs (if available)",
-        action="store_true",
+        "--tasks",
+        type=str,
+        nargs="+",
+        default=None,
+        help=(
+            "Space separated list of tasks to include in the ONNX model. "
+            "By default all global tasks are included.",
+        ),
     )
     parser.add_argument(
         "-mf",
         "--object_name",
+    )
+    parser.add_argument(
+        "--combine_outputs",
+        type=str,
+        nargs="+",
+        default=[],
+        help="Space seperated list of items described in 'parse_output_combination'",
+    )
+    parser.add_argument(
+        "-r",
+        "--rename",
+        type=str,
+        nargs="+",
+        default=[],
+        help="Space seperated list of them form 'oldname1:newname1 oldname2:newname2'",
     )
     parser.add_argument(
         "-f",
@@ -96,9 +115,11 @@ class ONNXModel(ModelWrapper):
         onnx_feature_map: list[dict],
         variable_map: dict,
         name: str | None = None,
-        include_aux: bool = False,
+        tasks_to_output: list[str] | None = None,
         object_name: str | None = None,
         mf_config: dict | None = None,
+        combine_outputs: list[tuple] | None = None,
+        rename_outputs: dict[str, str] | None = None,
         **kwargs,
     ) -> None:
         super().__init__(**kwargs)
@@ -107,9 +128,21 @@ class ONNXModel(ModelWrapper):
         assert "-" not in self.name, "Model name cannot contain dashes."
         for task in self.model.tasks:
             task.model_name = self.name
+        self.tasks_to_output = tasks_to_output or []
+        if not self.tasks_to_output:
+            self.tasks_to_output = [
+                t.name for t in self.model.tasks if t.input_name == self.global_object
+            ]
+            print("No tasks specified, dumping all global tasks: ", self.tasks_to_output)
+        else:
+            # Need aux tasks to be after global tasks, and in the following order for checks to work
+            for t in ["track_origin", "track_vertexing", "track_type"]:
+                if t in self.tasks_to_output:
+                    self.tasks_to_output.remove(t)
+                    self.tasks_to_output.append(t)
 
-        self.include_aux = include_aux
-
+        self.combine_outputs = combine_outputs or []
+        self.rename_outputs = rename_outputs or {}
         if sum([bool(object_name), bool(mf_config)]) not in {0, 2}:
             raise ValueError("If one of object name or mf config is defined, so must the other.")
         self.object = object_name
@@ -125,6 +158,7 @@ class ONNXModel(ModelWrapper):
         self.variable_map = variable_map
 
         example_input_list = []
+        num_tracks = 40
         self.salt_names = []
         self.input_names = []
         self.aux_sequence_index = 1
@@ -134,7 +168,7 @@ class ONNXModel(ModelWrapper):
                 example_input_list.append(torch.rand(1, self.input_dims[self.global_object]))
             else:
                 example_input_list.append(
-                    torch.rand(1, 40, self.input_dims[feature["name_salt"]]).squeeze(0)
+                    torch.rand(1, num_tracks, self.input_dims[feature["name_salt"]]).squeeze(0)
                 )
             if feature["name_salt"] == self.aux_sequence_object:
                 self.aux_sequence_index = i
@@ -149,18 +183,39 @@ class ONNXModel(ModelWrapper):
     @property
     def output_names(self) -> list[str]:
[…]
