### 2025-02-17 7f38f7f
Update training.md
