### 2025-02-24 371bc01
Add config file for GN2XTau
