+++ salt/configs/GN2XTau.yaml
@@ -1,0 +1,146 @@
+name: GN2XTauV00
+
+model:
+  lrs_config:
+    initial: 1e-7
+    max: 3e-4
+    end: 5e-5
+    pct_start: 0.01
+    weight_decay: 1e-5
+
+  model:
+    class_path: salt.models.SaltModel
+    init_args:
+      init_nets:
+        - input_name: tracks
+          dense_config:
+            output_size: &embed_dim 256
+            hidden_layers: [256]
+            activation: &activation ReLU
+        - input_name: flow
+          dense_config:
+            output_size: *embed_dim
+            hidden_layers: [256]
+            activation: *activation
+
+      encoder:
+        class_path: salt.models.TransformerV2
+        init_args:
+          num_layers: 8
+          embed_dim: *embed_dim
+          out_dim: &out_dim 128
+          attn_type: flash-varlen
+          norm: LayerNorm
+          dense_kwargs:
+            activation: *activation
+            dropout: 0.1
+            gated: True
+          attn_kwargs:
+            num_heads: 8
+            dropout: 0.1
+          num_registers: 8
+
+
+      pool_net:
+        class_path: salt.models.GlobalAttentionPooling
+        init_args: { input_size: *out_dim }
+
+      tasks:
+        class_path: torch.nn.ModuleList
+        init_args:
+          modules:
+            - class_path: salt.models.ClassificationTask
+              init_args:
+                name: jets_classification
+                input_name: jets
+                label: flavour_label
+                loss:
+                  class_path: torch.nn.CrossEntropyLoss
+                  init_args: { weight: [4.4, 1.5, 1.5, 2.8, 1.0]}
+                  #init_args: { weight: [0.0, 1.5, 1.5, 2.8, 1.0]}
+                dense_config: &task_dense_config
+                  input_size: *out_dim
+                  output_size: 5 #4
+                  hidden_layers: [128, 64, 32]
+                  activation: *activation
+
+            - class_path: salt.models.ClassificationTask
+              init_args:
+                name: track_origin
+                input_name: tracks
+                label: ftagTruthOriginLabel
+                weight: 0.33
+                loss:
+                  class_path: torch.nn.CrossEntropyLoss
+                  init_args:
+                    weight: [3.8, 163.7, 1.0, 15.8, 10.6, 11.7, 64.4, 22.4]
+                dense_config:
+                  <<: *task_dense_config
+                  output_size: 8
+                  context_size: *out_dim
+
+            - class_path: salt.models.VertexingTask
+              init_args:
+                name: track_vertexing
+                input_name: tracks
+                label: ftagTruthVertexIndex
+                weight: 0.75
+                loss:
+                  class_path: torch.nn.BCEWithLogitsLoss
+                  init_args: { reduction: none }
+                dense_config:
+                  <<: *task_dense_config
+                  input_size: 256
+                  output_size: 1
+                  context_size: *out_dim
+
+data:
+  num_inputs:
+    tracks: 100
+    flow: 100
+
+  variables:
+    jets:
+      - pt
+      - eta
+    tracks:
+      - d0
+      - z0SinTheta
+      - dphi
+      - deta
+      - qOverP
+      - lifetimeSignedD0Significance
+      - lifetimeSignedZ0SinThetaSignificance
+      - phiUncertainty
+      - thetaUncertainty
+      - qOverPUncertainty
+      - numberOfPixelHits
+      - numberOfSCTHits
[…]
