### 2025-03-05 c37476e
Refactor SaltCli.before_instantiate_classes
