+++ salt/utils/cli.py
@@ -7,6 +7,7 @@ import numpy as np
 import torch
 import yaml
 from ftag.git_check import check_for_uncommitted_changes, create_and_push_tag
+from jsonargparse import Namespace as JsonNamespace
 from jsonargparse.typing import register_type
 from lightning.pytorch.cli import LightningCLI
 
@@ -88,209 +89,216 @@ class SaltCLI(LightningCLI):
 
     def before_instantiate_classes(self) -> None:
         """A lot of automatic configuration is done here."""
-        sc = self.config[self.subcommand] if self.subcommand else self.config
+        config = self.config[self.subcommand] if self.subcommand else self.config
+        sc_tasks = config.model.model.init_args.tasks.init_args.modules
 
-        if "config_S3" in sc.data:
-            from salt.utils.file_utils import require_S3_CLI, setup_S3_CLI
-
-            if require_S3_CLI(sc.data["config_S3"]):
-                sc.data = setup_S3_CLI(sc.data)
-
-        # add the labels from the model config to the data config
         labels: dict = {}
-        model_dict = vars(sc.model.model.init_args)
-        for submodel in model_dict["tasks"]["init_args"]["modules"]:
-            assert "Task" in submodel["class_path"]
-            task = submodel["init_args"]
-            if task["input_name"] not in labels:
-                labels[task["input_name"]] = []
-            # Check if there is a merge dict and try if exists
-            if hasattr(self.config, "fit"):
-                fit_config = self.config.fit.model.model.init_args
-                if hasattr(fit_config, "merge_dict"):
-                    merge_dict = self.config.fit.model.model.init_args.merge_dict
-                    if merge_dict is not None:
-                        task_input_name = task["input_name"]
-                        if task_input_name in merge_dict:
-                            for inp in merge_dict[task["input_name"]]:
-                                if inp not in labels:
-                                    labels[inp] = []
-                                if self.subcommand == "fit":
-                                    if label := task.get("label"):
-                                        labels[inp].append(label)
-                                    if weight := task.get("sample_weight"):
-                                        labels[inp].append(weight)
-                                    if targets := task.get("targets"):
-                                        for target in listify(targets):
-                                            labels[inp].append(target)
-            if self.subcommand == "fit":
-                if label := task.get("label"):
-                    labels[task["input_name"]].append(label)
-                if weight := task.get("sample_weight"):
-                    labels[task["input_name"]].append(weight)
-                if targets := task.get("targets"):
-                    for target in listify(targets):
-                        labels[task["input_name"]].append(target)
-            if denominators := task.get("target_denominators"):
-                for denominator in listify(denominators):
-                    labels[task["input_name"]].append(denominator)
+        for task in sc_tasks:
+            assert "Task" in task["class_path"]
+            self.collect_labels_from_task(
+                task["init_args"], config.model.model.init_args.get("merge_dict"), labels
+            )
 
-        if model_dict.get("mask_decoder"):
-            if not (mf_config := sc["data"].get("mf_config")):
+        if config.model.model.init_args.get("mask_decoder"):
+            if not (maskformer_config := config.data.get("mf_config")):
                 raise ValueError("Mask decoder requires 'mf_config' in data config.")
-            if mf_config.constituent.name not in labels:
+            if maskformer_config.constituent.name not in labels:
                 raise ValueError(
-                    f"The constituent name {mf_config.constituent.name} is not in the data labels. "
-                    "Ensure that the constituent name is in the input_map of the data config."
+                    f"The constituent name {maskformer_config.constituent.name} is not in the"
[…]
