### 2025-03-07 59d0306
Fix bug in class_dict processing when remapping is used.
