### 2025-03-13 06f9274
fix mamba setup
