### 2025-04-07 2aa822b
Generalize Salt to make it more compatible for uses outside FTAG
