### 2025-04-08 efb6048
Automatically group train and val metrics in comet UI
