### 2025-04-15 312e08e
Configured the global_object name from input config automatically
