+++ salt/callbacks/predictionwriter.py
@@ -25,7 +25,6 @@ class PredictionWriter(Callback):
         half_precision: bool = False,
         object_classes: list | None = None,
         extra_vars: Vars | None = None,
-        global_object: str = "jets",
     ) -> None:
         """Write test outputs to h5 file.
 
@@ -51,9 +50,6 @@ class PredictionWriter(Callback):
         extra_vars : Vars
             Extra variables to write to file for each input type. If not specified for a given input
             type, all variables in the test file will be written.
-        global_object : str
-            Name of the global object the model is trained on (usually the target for
-            classification). By default "jets".
         """
         super().__init__()
         if extra_vars is None:
@@ -64,7 +60,6 @@ class PredictionWriter(Callback):
         self.half_precision = half_precision
         self.precision = "f2" if self.half_precision else "f4"
         self.object_classes = object_classes
-        self.global_object = global_object
 
     def setup(self, trainer: Trainer, module: LightningModule, stage: str) -> None:
         if stage != "test":
@@ -74,6 +69,7 @@ class PredictionWriter(Callback):
         # some basic info
         self.trainer = trainer
         self.ds = trainer.datamodule.test_dataloader().dataset
+        self.global_object = self.ds.global_object
         self.batch_size = trainer.datamodule.batch_size
         self.test_suff = trainer.datamodule.test_suff
         self.file = self.ds.file
