+++ salt/configs/Baseline_Xbb.yaml
@@ -22,7 +22,7 @@ trainer:
   callbacks:
     - class_path: salt.callbacks.Checkpoint
       init_args:
-        monitor_loss: val_jets_classification_loss
+        monitor_loss: val/jets_classification_loss
     - class_path: salt.callbacks.PredictionWriter
       init_args:
         jet_variables:
