### 2025-04-28 #7217 @GarrettMerz
Hi, all- apologies for the delay on this! I have a few comments on the software paper:

1) how are type A and type B inputs distinguished? Are these different types of reconstructed final-state particles (photons, muons, etc.) or different types of raw detector features (tracks, calo hits, etc.)? I'm assuming the latter, but this should be stated explicitly if so.

2) The line "the different types of constituents are then projected into a shared representation space by a series of initialisation networks. The embedded constituents are then combined and fed into a encoder network that processes constituents of different modalities in a unified way" is unclear/could use substantially more detail. Is the shared embedding space just a concatenation of two disjoint embedding spaces for each input type, or is it constructed some other way (i.e., using an autoencoder, a contrastive pretraining,or something else)? Is the init/encoder step constructed using a pretraining step on a dedicated pretraining dataset, or is it trained along with the downstream tasks? If possible, describing the model architecture as in a NeurIPS-style paper would be helpful for ML readers to understand the specifics of the training/design process- alternately, if this has been described elsewhere in more detail in in a different publication, it's fine to just cite that instead! But I'd like to read more about the specifics somewhere if I can :)

3) the GN tests take quite a while to run on my local NVidia GPU. Is this expected/ is there an issue with my local configuration (e.g., is it running the tests on CPU due to a misconfigured mamba environment on my end? Can/should you throw an error or a warning in the test suite if a GPU is expected but not found?)

Some miscellany-

1) The tutorial evidently cannot be run without access to lxplus or lvmfs. Similarly, the data files to run the pretraining etc. are hosted on the CERN ftag docs pages, which I cna no longer access. In order for this package to be published for a general audience rather than CERN users only, all CERN-gated information relevant to run the package should be either made public or replaced with publicly-accessible datasets, so non-CERN users (phenomenologists, ML scientists, etc.) can run and test the package.

2) Similarly, I get a "cannot find team" error when trying to join the Mattermost channel mentioned in the docs. Is this link correct? I am not currently affiliated with ATLAS/CMS (though I was in the recent past and may be again in the near future), so I am wondering if this error is due to my CERN user account no longer being active. If this is the case, please update this link or modify the docs so that non-CERN users can perform all actions mentioned in the documentation unless explicitly indicated otherwise.
