### 2025-04-28 #7217 @GarrettMerz
In general- this is clearly a well-made software package that fits neatly into the current ML-HEP conversation, so I encourage publication of this package and paper- however, some additional effort must be done to make the work sufficiently accessible to the general JOSS readership.
