### 2025-05-05 4fcfe3c
Add electrons to GN3v00 config
