+++ salt/configs/GN3v01/GN3V00.yaml
@@ -1,0 +1,181 @@
+# Baseline GN3V00 model
+name: GN3V00
+
+model:
+  # Lion GLS addition
+  optimizer: lion
+  loss_mode: GLS
+  lrs_config:
+    initial: 2e-8
+    max: 1e-4
+    end: 2e-6
+    pct_start: 0.01
+    weight_decay: 1e-5
+
+  model:
+    class_path: salt.models.SaltModel
+    init_args:
+      init_nets:
+        - input_name: tracks
+          dense_config: &init
+            output_size: &embed_dim 512
+            hidden_layers: [512]
+            activation: &activation SiLU
+        # Adding flows
+        - input_name: flows
+          dense_config:
+            <<: *init
+
+      encoder:
+        class_path: salt.models.TransformerV2
+        init_args:
+          num_layers: 4
+          embed_dim: *embed_dim
+          out_dim: &out_dim 256
+          # flash attention
+          attn_type: flash-varlen
+          norm: LayerNorm
+          dense_kwargs:
+            activation: *activation
+            gated: True
+          attn_kwargs:
+            num_heads: 8
+          num_registers: 8
+
+      pool_net:
+        class_path: salt.models.GlobalAttentionPooling
+        init_args: { input_size: *out_dim }
+
+      tasks:
+        class_path: torch.nn.ModuleList
+        init_args:
+          modules:
+            - class_path: salt.models.ClassificationTask
+              init_args:
+                name: jets_classification
+                input_name: jets
+                label: flavour_label
+                use_class_dict: True
+                loss: torch.nn.CrossEntropyLoss
+                dense_config: &task_dense_config
+                  input_size: *out_dim
+                  # split light classes
+                  output_size: 6
+                  hidden_layers: [128, 64, 32]
+                  activation: *activation
+            - class_path: salt.models.ClassificationTask
+              init_args:
+                name: track_origin
+                input_name: tracks
+                label: ftagTruthOriginLabel
+                use_class_dict: True
+                loss: torch.nn.CrossEntropyLoss
+                dense_config:
+                  <<: *task_dense_config
+                  output_size: 8
+                  context_size: *out_dim
+                  activation: *activation
+            - class_path: salt.models.VertexingTask
+              init_args:
+                name: track_vertexing
+                input_name: tracks
+                label: ftagTruthVertexIndex
+                loss:
+                  class_path: torch.nn.BCEWithLogitsLoss
+                  init_args: { reduction: none }
+                dense_config:
+                  <<: *task_dense_config
+                  input_size: 512
+                  output_size: 1
+                  context_size: *out_dim
+                  activation: *activation
+            # TrackType task
+            - class_path: salt.models.ClassificationTask
+              init_args:
+                name: track_type
+                input_name: tracks
+                label: ftagTruthTypeLabel
+                label_map: { -2: 2, -3: 3, 5: 4, -5: 4, 6: 5, -6: 5 }
+                class_names: [NoTruth, Other, Pion, Kaon, Electron, Muon]
+                loss:
+                  class_path: torch.nn.CrossEntropyLoss
+                  init_args:
+                    weight: [3.5, 13.34, 1.0, 6.47, 21.5, 112.1]
+                dense_config:
+                  <<: *task_dense_config
+                  output_size: 6
+                  context_size: *out_dim
+                  activation: *activation
+            # Jet pT regression task
[…]
