+++ salt/configs/GN3v01/GN3_SoftE.yaml
@@ -1,0 +1,94 @@
+# On top of GN3_Hybrid or GN3_Charge
+# Adds electrons
+name: GN3_SoftE
+
+model:
+  model:
+    class_path: salt.models.SaltModel
+    init_args:
+      init_nets:
+        - input_name: tracks
+          dense_config: &init
+            output_size: &embed_dim 512
+            hidden_layers: [512]
+            activation: &activation SiLU
+        # Adding flows
+        - input_name: flows
+          dense_config:
+            <<: *init
+        # Adding electrons
+        - input_name: electrons
+          dense_config:
+            <<: *init
+
+data:
+  input_map:
+    jets: jets
+    tracks: tracks_ghost
+    flows: flows
+    electrons: electrons
+  variables:
+    jets:
+      - pt_btagJes
+      - eta_btagJes
+    tracks:
+      - d0
+      - z0SinTheta
+      - dphi
+      - deta
+      - qOverP
+      - IP3D_signed_d0_significance
+      - IP3D_signed_z0_significance
+      - phiUncertainty
+      - thetaUncertainty
+      - qOverPUncertainty
+      - numberOfPixelHits
+      - numberOfSCTHits
+      - numberOfInnermostPixelLayerHits
+      - numberOfNextToInnermostPixelLayerHits
+      - numberOfInnermostPixelLayerSharedHits
+      - numberOfInnermostPixelLayerSplitHits
+      - numberOfPixelSharedHits
+      - numberOfPixelSplitHits
+      - numberOfSCTSharedHits
+      # Soft Muon variables
+      - leptonID
+      - muon_quality
+      - muon_qOverPratio
+      - muon_momentumBalanceSignificance
+      - muon_scatteringNeighbourSignificance
+    flows:
+      - pt
+      - energy
+      - deta
+      - dphi
+      - isCharged
+    electrons:
+      - pt
+      - ptfrac
+      - ptrel
+      - dr
+      - abs_eta
+      - eta
+      - phi
+      - ftag_et
+      - qOverP
+      - d0RelativeToBeamspot
+      - d0RelativeToBeamspotSignificance
+      - ftag_ptVarCone30OverPt
+      - numberOfPixelHits
+      - numberOfSCTHitsInclDead
+      - ftag_deltaPOverP
+      - eProbabilityHT
+      - deltaEta1
+      - deltaPhiRescaled2
+      - ftag_energyOverP
+      - Rhad
+      - Rhad1
+      - Eratio
+      - weta2
+      - Rphi
+      - Reta
+      - wtots1
+      - f1
+      - f3
