### 2025-05-10 2cc8ac7
tweak class weights for truthtype pred task for new sample
