+++ salt/configs/GN3v01/GN3_Charge.yaml
@@ -88,7 +88,7 @@ model:
                 loss:
                   class_path: torch.nn.CrossEntropyLoss
                   init_args:
-                    weight: [3.5, 13.34, 1.0, 6.47, 21.5, 112.1]
+                    weight: [3.56, 13.3, 1.0, 6.49, 21.48, 109.72]
                 dense_config:
                   <<: *task_dense_config
                   output_size: 6
