### 2025-05-13 247d9ff
Fix in classes weights and name for jet-charge aux task
