### 2025-05-15 3c3fb6b
JOSS paper & documentation updates
