+++ mkdocs.yaml
@@ -37,8 +37,8 @@ nav:
   - Evaluation: evaluation.md
   - ONNX Export: export.md
   - Contributing: contributing.md
-  - Tutorial: tutorial.md
-  - Tutorial (Xbb): tutorial-Xbb.md
+  - Tutorial (public): tutorial.md
+  - Tutorial (internal): tutorial-Xbb.md
   - API Reference:
       - api/data.md
       - api/initialisation.md
