+++ salt/callbacks/__init__.py
@@ -1,5 +1,6 @@
 from salt.callbacks.checkpoint import Checkpoint
 from salt.callbacks.gradient_logger import GradientLoggerCallback
+from salt.callbacks.integrated_gradients_writer import IntegratedGradientWriter
 from salt.callbacks.maskformer_metrics import MaskformerMetrics
 from salt.callbacks.performancewriter import PerformanceWriter
 from salt.callbacks.predictionwriter import PredictionWriter
@@ -9,6 +10,7 @@ from salt.callbacks.weight_logger import WeightLoggerCallback
 __all__ = [
     "Checkpoint",
     "GradientLoggerCallback",
+    "IntegratedGradientWriter",
     "MaskformerMetrics",
     "PerformanceWriter",
     "PredictionWriter",
