+++ requirements-ig.txt
@@ -1,0 +1,1 @@
+git+ssh://git@gitlab.cern.ch:7999/atlas-flavourtagging-algorithms/salt-attribution.git@master#egg=salt-attribution
