+++ salt/tests/test_datasets.py
@@ -1,3 +1,4 @@
+import h5py
 import torch
 from ftag import get_mock_file
 
@@ -25,3 +26,112 @@ def test_track_selector():
         assert torch.any(inputs["tracks"][~masks["tracks"]] > 3.5)
         inputs, masks, _ = sel[i : i + 10]
         assert not torch.any(inputs["tracks"][~masks["tracks"]] > 3.5)
+
+
+def test_input_batch():
+    f = get_mock_file()[0]
+    norm_dict = {}
+    variables = {
+        "jets": [
+            "pt",
+            "eta",
+            "R10TruthLabel_R22v1",
+            "GhostBHadronsFinalCount",
+            "GhostCHadronsFinalCount",
+        ],
+        "tracks": ["d0"],
+    }
+    labels = {"jets": ["flavour_label"], "tracks": ["numberOfSCTSharedHits"]}
+    cn = ["hbb", "hcc", "top", "qcdbb", "qcdbx", "qcdcx", "qcdll"]
+    dataset = SaltDataset(
+        f, norm_dict, variables, "train", labels=labels, use_labeller=False, class_names=cn
+    )
+    input_map = {k: k for k in variables}
+    for i in range(0, len(dataset), 10):
+        ds = dataset[i : i + 10]
+        labels = ds[2]
+        for input_name in input_map:
+            batch = dataset.arrays[input_name]
+            labels[input_name] = {}
+        assert batch.shape == (10, 40)
+        assert input_map == {"jets": "jets", "tracks": "tracks"}
+
+
+def test_process_labels():
+    f = get_mock_file()[0]
+    norm_dict = {}
+    variables = {
+        "jets": [
+            "pt",
+            "eta",
+            "R10TruthLabel_R22v1",
+            "GhostBHadronsFinalCount",
+            "GhostCHadronsFinalCount",
+        ]
+    }
+    labels = {"jets": ["flavour_label"]}
+    cn = ["hbb", "hcc", "top", "qcdbb", "qcdbx", "qcdcx", "qcdll"]
+    dataset = SaltDataset(
+        f, norm_dict, variables, "train", labels=labels, use_labeller=True, class_names=cn
+    )
+    input_map = {k: k for k in variables}
+    for i in range(0, len(dataset), 10):
+        ds2 = dataset[i : i + 10]
+        labels = ds2[2]
+        for input_name in input_map:
+            batch = dataset.arrays[input_name]
+            labels[input_name] = {}
+            output = dataset.process_labels(labels, batch, input_name)
+            assert isinstance(*output.values(), torch.Tensor) is True
+
+
+def test_file_vars_for_cuts():
+    f = get_mock_file()[0]
+    norm_dict = {}
+    file_open = h5py.File(f, "r")
+    input_variables = {"jets": ["pt", "eta", "R10TruthLabel_R22v1", "GhostCHadronsFinalCount"]}
+    all_file_vars = list(file_open["jets"].dtype.fields.keys())
+    labels = {"jets": ["flavour_label"]}
+    cn = ["hbb", "hcc", "top", "qcdbb", "qcdbx", "qcdcx", "qcdll"]
+    dataset = SaltDataset(
+        f, norm_dict, input_variables, "train", labels=labels, use_labeller=True, class_names=cn
+    )
+    all_unique_cuts = list(
+        set(sum((label.cuts.variables for label in dataset.labeller.labels), []))
+    )
+    assert not (all(x in input_variables["jets"] for x in all_unique_cuts))
+    assert all(x in all_file_vars for x in all_unique_cuts)
+
+
+def test_process_labels_size_check():
+    f = get_mock_file()[0]
+    norm_dict = {}
+    variables = {
+        "jets": [
+            "pt",
+            "eta",
[…]
