+++ salt/tests/test_edges.py
@@ -38,7 +38,7 @@ def test_transformer_edges() -> None:
     assert not torch.isnan(out).any()
 
     # test that adding a padded track does not change the output
-    """tracks = torch.rand(1, 10, 10)
+    tracks = torch.rand(1, 10, 10)
     edges = torch.rand(1, 10, 10, 10)
     mask = get_random_mask(1, 10, p_valid=1)
     out = net(tracks, edges, pad_mask=mask)
@@ -48,7 +48,8 @@ def test_transformer_edges() -> None:
     mask = torch.zeros(tracks.shape[:-1]).bool()
     mask[:, -1] = True
     out_with_pad = net(tracks, edges, pad_mask=mask)[:, :-1]
-    assert torch.all(out == out_with_pad)"""
+    tensor_check = out.data.half() == out_with_pad.data.half()
+    assert torch.all(tensor_check)
 
 
 def test_mha_edges_allvalid_mask() -> None:
