+++ salt/tests/test_models.py
@@ -73,8 +73,7 @@ def test_transformer() -> None:
     # test fully padded case
     out = net(torch.rand(1, 10, 10), pad_mask=get_random_mask(1, 10, p_valid=0))
     assert not torch.isnan(out).any()
-    """
-    disabling for now as it's failing in the CI
+
     # test that adding a padded track does not change the output
     tracks = torch.rand(1, 10, 10)
     mask = get_random_mask(1, 10, p_valid=1)
@@ -82,9 +81,9 @@ def test_transformer() -> None:
     tracks = torch.cat([tracks, torch.zeros((1, 1, tracks.shape[2]))], dim=1)
     mask = torch.zeros(tracks.shape[:-1]).bool()
     mask[:, -1] = True
-    assert torch.all(out.data == out_with_pad.data)
     out_with_pad = net(tracks, pad_mask=mask)[:, :-1]
-    """
+    tensor_check = out.data.half() == out_with_pad.data.half()
+    assert torch.all(tensor_check)
 
 
 def test_transformer_cross_attention_encoder() -> None:
