### 2025-06-06 f2b7f24
Add require_labels flag in relabelling
