### 2025-06-10 0bcc6b8
Fix GN2 config
