### 2025-06-17 c0db1f3
Map malformed truthOriginLabel values as invalid tracks
