+++ requirements.txt
@@ -16,6 +16,6 @@ jsonargparse[signatures,argcomplete]==4.32.0
 torchmetrics==1.2.1
 onnx==1.16.0
 onnxruntime==1.15.1
-atlas-ftag-tools==0.2.6
+atlas-ftag-tools==0.2.8
 scipy==1.12.0
 lion-pytorch==0.2.2
