### 2025-07-02 2cf8b32
Remove maskformer dependence from object-level tasks
