### 2025-07-15 #7217 @arfon
@GarrettMerz – thanks for the quick response. Could I ask you to update your [review checklist](https://github.com/openjournals/joss-reviews/issues/7217#issuecomment-2773375818) too please?
