### 2025-07-23 #7217 @samvanstroud
Gentle reminder @GarrettMerz (or check @arfon if we can proceed already?) We have a few papers in the final stages of peer review and would love to be able to cite the JOSS DOI :)
