+++ salt/models/maskformer.py
@@ -54,6 +54,8 @@ class MaskDecoder(nn.Module):
         num_objects: int,
         loss_config: Mapping,
         aux_loss: bool = False,
+        constituent_name=None,
+        class_weights: list[float] | None = None,
     ):
         super().__init__()
         self.aux_loss = aux_loss
@@ -71,7 +73,8 @@ class MaskDecoder(nn.Module):
         self.norm1 = nn.LayerNorm(embed_dim)
         self.norm2 = nn.LayerNorm(embed_dim)
 
-        self.mask_loss = MaskFormerLoss(**loss_config, num_objects=num_objects)
+        self.mask_loss = MaskFormerLoss(**loss_config, num_objects=num_objects, class_weights=class_weights)
+        self.constituent_name = constituent_name
 
     def get_preds(
         self,
@@ -166,14 +169,36 @@ class MaskDecoder(nn.Module):
         # MF only supports one input, if we have multiple then we have no way of knowing
         # what section of the embedding relates to objects we want to generate masks for
         if isinstance(pad_mask, dict):
-            assert len(pad_mask) == 1, "Maskformer only supports one input."
-            pad_mask = next(iter(pad_mask.values()))
+            if self.constituent_name is not None:
+                # Create the mask to extract only the relevent consituent embeddings
+                emb_mask = torch.zeros(preds["embed_xs"].shape[1], dtype=torch.bool, device=preds["embed_xs"].device)
+
+                i = 0
+                for k, v in pad_mask.items():
+                    if k == self.constituent_name:
+                        
+                        emb_mask[i : i + v.shape[1]] = True
+                        break
+                    i += v.shape[1]
+                pad_mask = pad_mask[self.constituent_name]
+            else:
+                assert len(pad_mask) == 1, "Maskformer only supports one input."
+                pad_mask = next(iter(pad_mask.values()))
+                emb_mask = torch.ones(preds["embed_xs"].shape[1], dtype=torch.bool, device=preds["embed_xs"].device)
+        else:
+            emb_mask = torch.ones(preds["embed_xs"].shape[1], dtype=torch.bool, device=preds["embed_xs"].device)
 
         x = preds["embed_xs"]
+        # print("original x shape", x.shape, "pad_mask shape", pad_mask.shape, emb_mask.shape)
+        x = x[:, emb_mask, :]
+        # print("Next x shape", x.shape, "pad_mask shape", pad_mask.shape)
         # apply norm
         q = self.norm1(self.inital_q.expand(x.shape[0], -1, -1))
         x = self.norm2(x)
-
+        # print("Initial q shape", q.shape, "x shape", x.shape)
+        # print(q[0])
+        # print(q[1])
+        # print("Pad mask?", pad_mask[0], pad_mask[1])
         # Add a dummy track to the inputs (and to pad mask) to stop onnx complaining
         xpad = torch.zeros((x.shape[0], 1, x.shape[-1]), device=x.device, dtype=x.dtype)
         x = torch.cat([x, xpad], dim=1)
@@ -184,13 +209,22 @@ class MaskDecoder(nn.Module):
             pad_mask = torch.cat([pad_mask, padpad_mask], dim=1)
 
         intermediate_outputs: list | None = [] if self.aux_loss else None
+        # print("Weird before? ", q[0])
+        # print("X before? ", x[0])
+        i=0
         for layer in self.layers:
+            # print(i, "Weird still,", q[0][0], q[1][0], x[0][0], x[1][0])
             if self.aux_loss:
                 assert intermediate_outputs is not None
[…]
