### 2025-08-04 f561035
final paper typos
