+++ paper/paper.md
@@ -97,7 +97,7 @@ Some key features of `Salt` are listed below:
 - Multimodal, multitask models: `Salt` models support multimodal inputs and can be configured to perform various tasks such as classification, regression, segmentation, and edge classification tasks. Any combination of these can be used to flexibly define models for multitask learning problems.
 - Customisable and extensible: `Salt` supports full customisation of training and model configuration through YAML config files. Its modular design allows for the easy integration of custom dataloaders, layers, and models.
 - Train at scale: `Salt` can handle large volumes of data with efficient HDF5 [@hdf5:2023] dataloaders. It also includes multi-GPU support from Lightning, enabling distributed training.
-- Deployment ready: `Salt` facilitates ONNX [@onnx] serialization for integrating models into C++ based software environments.
+- Deployment ready: `Salt` facilitates ONNX [@onnx] serialisation for integrating models into C++ based software environments.
 
 
 # Statement of need
@@ -119,7 +119,7 @@ This architecture facilitates the training of multimodal and multitask models as
 In the context of jet classification, these input modalities might include global features of the jet and varying numbers of jet constituents such as charged particle trajectories, calorimeter energy depositions, reconstructed leptons, or inner detector spacepoints.
 The architecture is described briefly below.
 First, each input type (e.g., tracks, calorimeter deposits) is independently projected into a common embedding space of fixed dimension using separate initialisation networks.
-These initialisation networks can optionally concatenate global features with constituent features and apply positional encoding to certain features (for example azimuthal angle).
+These initialisation networks can optionally concatenate global features with constituent features and apply positional encoding to certain features, for example azimuthal angle.
 Once embedded, the different types of constituents are considered to be in the same semantic space and are processed together by a transformer encoder that allows them to interact through stacked multi-head attention layers.
 The encoder maintains the same embedding dimension throughout its layers and can optionally update edge features if they are present.
 The encoder then outputs to a set of task-specific modules, each tailored to a specific learning objective.
@@ -127,7 +127,7 @@ The initialisation networks, transformer encoder, and task-specific networks are
 This approach allows the model to leverage all the available detector information, leading to improved performance.
 Concrete examples of this architecture are in use at ATLAS [@GN1; @GN2X; @GN2].
 
-![This diagram illustrates the flow of information within a generic model trained using `Salt`. In this example, global object features are provided alongisde two types of constituents, "Type A" and "Type B", which represent different input modalities such as charged particle trajectories or calorimeter energy depositions. The model is configured with three training objectives, each of which may relate to the global object or the one of the constituent modalities. Concatenation is denoted by $\oplus$.\label{fig:salt-arch}](salt-arch.png){ width=90% }
+![This diagram illustrates the flow of information within a generic model trained using `Salt`. In this example, global object features are provided alongside two types of constituents, "Type A" and "Type B", which represent different input modalities such as charged particle trajectories or calorimeter energy depositions. The model is configured with three training objectives, each of which may relate to the global object or one of the constituent modalities. Concatenation is denoted by $\oplus$.\label{fig:salt-arch}](salt-arch.png){ width=90% }
 
 
 # Related work
