### 2025-08-19 9e1dd92
remove mandatory maskformer for object predictionwriter
